majit-trace: compute the stats op-reduction percentage in f64
`Logger::summary` computed the op-count reduction as
`(total_before - total_after)` on usize, which underflows and panics in
debug builds under MAJIT_STATS when optimization raises the op count
(`total_after > total_before`). Compute the difference in f64 so a
grown-op-count trace reports a negative reduction instead of aborting the
summary on drop.

+++ majit/majit-trace/src/logger.rs
@@ -183,8 +183,11 @@ impl Logger {
     pub fn summary(&self) -> String {
         let total_before = self.total_ops_before();
         let total_after = self.total_ops_after();
+        // Signed in f64: optimization can raise the op count (guards/spills), so
+        // `total_after > total_before` is a legitimate negative reduction, not an
+        // unsigned-subtraction underflow.
         let reduction_pct = if total_before > 0 {
-            ((total_before - total_after) as f64 / total_before as f64) * 100.0
+            (total_before as f64 - total_after as f64) / total_before as f64 * 100.0
         } else {
             0.0
         };
